backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Vector Database
chromadb==0.5.2

# LLM & Embeddings
openai==1.50.0
tenacity==8.2.3

# Document Processing
pypdf==4.0.1
python-docx==1.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
